code cleanup

+++ src/modules/commands.ts
@@ -12,51 +12,32 @@ export const d: Command = {
   reqMod: true,
 
   run: async (guild, msg) => {
-    const match = /(!d)\s(\w+)\s(.+)/g.exec(msg.content)
 
-    if (!match) {
-      await msg.channel.send("Unable to match Discord Tag.").catch(console.error)
-      return
-    }
+    const match = /(!d)\s(\w+)\s(.+)/g.exec(msg.content)
+    if (!match) return await msg.channel.send("Invalid !d command. (e.g. !d user#1234 reason)").catch(console.error)
 
     const name = match[2].toLowerCase()
 
     const applicant = await getApplicant(name)
-    if (!applicant) {
-      await msg.channel.send(`Applicant does not exist: ${name}`).catch(console.error)
-      return
-    }
+    if (!applicant) return await msg.channel.send(`Applicant does not exist: ${name}`).catch(console.error)
 
     const channel = guild.channels.resolve(applicant.channelID)
-    if (!channel) {
-      await msg.channel
-        .send(`Channel does not exist for applicant: ${applicant.tag}`)
-        .catch(console.error)
-      return
-    }
-
-    if (!isTextChannel(channel)) {
-      await msg.channel.send(`Channel for applicant is not a text channel.`).catch(console.error)
-      return
-    }
+    if (!channel) return await msg.channel.send(`Channel does not exist for applicant: ${applicant.tag}`).catch(console.error)
+    if (!isTextChannel(channel)) return await msg.channel.send(`Channel for applicant is not a text channel.`).catch(console.error)
 
     channel
       .send(
-        `<@${applicant.memberID}>\n\n${match[3]}\n\nPlease type \`!confirm\` to acknowledge that you have read this message. Upon confirmation your application will be closed and you will be removed from the server.`
+        `<@${applicant.memberID}>\n\n${match[3]}\n\nPlease click the 👍 reaction on this message to confirm that you have read this message. Upon confirmation your application will be closed and you will be removed from the server.`
       )
       .catch(console.error)
 
     const appsChannel = guild.channels.resolve(channelCache.getOrThrow("apps").id)
-    if (!appsChannel) {
-      throw Error(`channel does not exist`)
-    }
-    if (!isTextChannel(appsChannel)) {
-      throw Error(`apps channel is not a Text Channel | ${appsChannel?.id}`)
-    }
+    if (!appsChannel) throw Error(`channel does not exist`)
+    if (!isTextChannel(appsChannel)) throw Error(`apps channel is not a text channel | ${appsChannel?.id}`)
 
     const appMessage = appsChannel.messages.resolve(applicant.appMessageID)
 
-    appMessage?.react(emojiCache.getOrThrow("declined").id)
+    appMessage?.react(emojiCache.getOrThrow("declined").id).catch(console.error)
   },
 }
 
@@ -65,26 +46,18 @@ export const confirm: Command = {
   reqMod: false,
 
   run: async (guild, msg) => {
-    if (!isTextChannel(msg.channel)) {
-      throw Error(`channel is not a text channel, got type ${msg.channel.type}`)
-    }
 
-    if (msg.channel.parent?.id != channelCache.getOrThrow("applicants").id) {
-      msg.channel.send(`This channel is not an applicant channel.`)
-      return
-    }
+    if (!isTextChannel(msg.channel)) throw Error(`channel is not a text channel, got type ${msg.channel.type}`)
+
+    if (msg.channel.parent?.id != channelCache.getOrThrow("applicants").id) return msg.channel.send(`This channel is not an applicant channel.`).catch(console.error)
 
     await msg.channel.delete().catch(console.error)
 
     const applicant = await getApplicant(msg.channel.name)
-    if (!applicant || !applicant.memberID) {
-      throw Error(`applicant does not exist: ${msg.channel.name}`)
-    }
+    if (!applicant || !applicant.memberID) throw Error(`applicant does not exist: ${msg.channel.name}`)
 
     const member = guild.members.resolve(applicant.memberID)
-    if (!member) {
-      throw Error(`member does not exist: ${applicant.tag} | ${applicant.memberID}`)
-    }
+    if (!member) throw Error(`member does not exist: ${applicant.tag} | ${applicant.memberID}`)
 
     await member.kick().catch(console.error)
 
@@ -96,45 +69,29 @@ export const a: Command = {
   reqMod: true,
 
   run: async (guild, msg) => {
+
     const match = /(!a)\s(\w+)/g.exec(msg.content)
 
-    if (!match) {
-      await msg.channel.send("Unable to match Discord Tag.").catch(console.error)
-      return
-    }
+    if (!match) return await msg.channel.send("Invalid !a command. (e.g !a user#1234)").catch(console.error)
 
     const name = match[2].toLowerCase()
 
     const applicant = await getApplicant(name)
-    if (!applicant || !applicant.memberID) {
-      await msg.channel.send(`Applicant does not exist: ${name}`).catch(console.error)
-      return
-    }
+    if (!applicant || !applicant.memberID) return await msg.channel.send(`Applicant does not exist: ${name}`).catch(console.error)
 
     const member = guild.members.resolve(applicant.memberID)
-    if (!member) {
-      throw Error(`member does not exist: ${applicant.tag} | ${applicant.memberID}`)
-    }
+    if (!member) throw Error(`member does not exist: ${applicant.tag} | ${applicant.memberID}`)
 
     member.roles.remove(roleCache.getOrThrow("Applicant").id)
 
     const channel = guild.channels.resolve(applicant.channelID)
-    if (!channel) {
-      await msg.channel
-        .send(`Channel does not exist for applicant: ${applicant.tag}`)
-        .catch(console.error)
-      return
-    }
+    if (!channel) return await msg.channel.send(`Channel does not exist for applicant: ${applicant.tag}`).catch(console.error)
 
     await channel.delete().catch(console.error)
 
     const appsChannel = guild.channels.resolve(channelCache.getOrThrow("apps").id)
-    if (!appsChannel) {
-      throw Error(`channel does not exist`)
-    }
-    if (!isTextChannel(appsChannel)) {
-      throw Error(`apps channel is not a Text Channel | ${appsChannel?.id}`)
-    }
+    if (!appsChannel) throw Error(`channel does not exist`)
+    if (!isTextChannel(appsChannel)) throw Error(`apps channel is not a Text Channel | ${appsChannel?.id}`)
 
     const appMessage = appsChannel.messages.resolve(applicant.appMessageID)
 
@@ -142,4 +99,12 @@ export const a: Command = {
 
     await removeApplicant(applicant)
   },
+}
+
+export const oc: Command = {
+  reqMod: false,
+
+  run: async (guild, msg) => {
+    msg.channel.send("https://cdn.discordapp.com/attachments/421529280355237891/757685187281944636/overlmao.png").catch(console.error)
+  }
 }
